Fix login page import, fixtures for home page
