+++ tests/e2e/fixtures/base.ts
@@ -1,5 +1,7 @@
 import { test as base } from "@playwright/test";
-import { AdminHomePage, LoginPage, StateHomePage } from "../pages/common";
+import LoginPage from "../pages/login.page";
+import StateHomePage from "../pages/stateHome.page";
+import AdminHomePage from "../pages/adminHome.page";
 
 type CustomFixtures = {
   loginPage: LoginPage;
@@ -7,6 +9,7 @@ type CustomFixtures = {
   adminHomePage: AdminHomePage;
 };
 
+// This new "test" can be used in multiple test files, and each of them will get the fixtures.
 export const test = base.extend<CustomFixtures>({
   loginPage: async ({ page }, use) => {
     await use(new LoginPage(page));
